fix: clear complexity gate regressions
