+++ packages/core/src/waste-management-output.render.ts
@@ -1,59 +1,15 @@
 import type { WasteCalendarPdfDocument } from './waste-management-output.types.js';
 import { PdfBuilder } from './waste-management-output.pdf-builder.js';
-
-type RgbColor = readonly [red: number, green: number, blue: number];
+import {
+  abbreviateHolidayLabel,
+  getEntryLabelWidth,
+  pad2,
+  splitLegendLabel,
+  type RgbColor,
+} from './waste-management-output.render.helpers.js';
 
 const PAGE_WIDTH = 841.89;
 const PAGE_HEIGHT = 595.28;
-
-const splitLegendLabel = (label: string): readonly string[] => {
-  if (label.length <= 24) {
-    return [label];
-  }
-
-  const words = label.split(/\s+/);
-  const lines: string[] = [];
-  let current = '';
-
-  for (const word of words) {
-    const next = current.length === 0 ? word : `${current} ${word}`;
-    if (next.length > 24 && current.length > 0) {
-      lines.push(current);
-      current = word;
-      continue;
-    }
-    current = next;
-  }
-
-  if (current.length > 0) {
-    lines.push(current);
-  }
-
-  return lines;
-};
-
-const pad2 = (value: number): string => value.toString().padStart(2, '0');
-
-const abbreviateHolidayLabel = (label: string): string => {
-  switch (label) {
-    case 'Christi Himmelfahrt':
-      return 'Christi Himmelf.';
-    case 'Tag der Deutschen Einheit':
-      return 'Tag d. Dt. Einheit';
-    default:
-      return label;
-  }
-};
-
-const getEntryLabelWidth = (code: string): number => {
-  if (code.length <= 2) {
-    return 18;
-  }
-  if (code.length === 3) {
-    return 22;
-  }
-  return 26;
-};
 
 type TextDrawInput = Readonly<{
   color?: RgbColor;
